% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffregr.R
\name{beta.mat.diffregr}
\alias{beta.mat.diffregr}
\title{Computation beta matrix}
\usage{
beta.mat.diffregr(ind1, ind2, beta1, beta2, beta, sig1, sig2, sig, Sig)
}
\arguments{
\item{ind1}{no descr}

\item{ind2}{no descr}

\item{beta1}{no descr}

\item{beta2}{no descr}

\item{beta}{no descr}

\item{sig1}{no descr}

\item{sig2}{no descr}

\item{sig}{no descr}

\item{Sig}{no descr}
}
\value{
no descr
}
\description{
Computation beta matrix
}
\author{
n.stadler
}
\keyword{internal}

