% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffnet.R
\name{bic.glasso}
\alias{bic.glasso}
\title{BIC.glasso}
\usage{
bic.glasso(x, lambda, penalize.diagonal = FALSE, plot.it = TRUE,
  use.package = "huge", include.mean = FALSE)
}
\arguments{
\item{x}{no descr}

\item{lambda}{no descr}

\item{penalize.diagonal}{no descr}

\item{plot.it}{no descr}

\item{use.package}{no descr}

\item{include.mean}{no descr}
}
\value{
no descr
}
\description{
BIC.glasso
}
\author{
n.stadler
}
\keyword{internal}

