% Generated by roxygen2 (4.0.2): do not edit by hand
\name{buildDotPlotDataFrame}
\alias{buildDotPlotDataFrame}
\title{Build up dataframe for plotting dot plot with ggplot2}
\usage{
buildDotPlotDataFrame(net.clustering, cluster.names, node.names)
}
\arguments{
\item{net.clustering}{Clustering}

\item{cluster.names}{Cluster names}

\item{node.names}{Node names}
}
\value{
A data frame for plotting the dotPlot with ggplot2 is returned.
Column P.Corr contains the partial correlations of each edge as a numeric,
column Mean contains the minimum mean expression of the two proteins
(e.g. if the edge is e(p1, p2), then the column contains min(mean(p1),
mean(p2))), column Edge contains the name of the edge as a character
string of the form "p1-p2" and column Type contains the cluster name of the
cluster that the edge belongs to as a character string.
}
\description{
Internal function
}
\keyword{internal}

