% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cv.glasso}
\alias{cv.glasso}
\title{Crossvalidation for GLasso}
\usage{
\method{cv}{glasso}(x, folds = 10, lambda, penalize.diagonal = FALSE,
  plot.it = FALSE, se = TRUE, include.mean = FALSE)
}
\arguments{
\item{x}{no descr}

\item{folds}{no descr}

\item{lambda}{lambda-grid (increasing!)}

\item{penalize.diagonal}{no descr}

\item{plot.it}{no descr}

\item{se}{no descr}

\item{include.mean}{no descr}
}
\value{
no descr
}
\description{
Crossvalidation for GLasso
}
\details{
8! lambda-grid has to be increasing (see glassopath)
}
\author{
n.stadler
}
\keyword{internal}

