% Generated by roxygen2 (4.0.2): do not edit by hand
\name{diffnet_pval}
\alias{diffnet_pval}
\title{P-value calculation}
\usage{
diffnet_pval(x1, x2, x, sig1, sig2, sig, mu1, mu2, mu, act1, act2, act,
  compute.evals, include.mean, method.compquadform, acc, epsabs, epsrel,
  show.warn)
}
\arguments{
\item{x1}{no descr}

\item{x2}{no descr}

\item{x}{no descr}

\item{sig1}{no descr}

\item{sig2}{no descr}

\item{sig}{no descr}

\item{mu1}{no descr}

\item{mu2}{no descr}

\item{mu}{no descr}

\item{act1}{no descr}

\item{act2}{no descr}

\item{act}{no descr}

\item{compute.evals}{no descr}

\item{include.mean}{no descr}

\item{method.compquadform}{no descr}

\item{acc}{no descr}

\item{epsabs}{no descr}

\item{epsrel}{no descr}

\item{show.warn}{no descr}
}
\value{
no descr
}
\description{
P-value calculation
}
\author{
n.stadler
}
\keyword{internal}

