% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffregr.R
\name{diffregr_pval}
\alias{diffregr_pval}
\title{Computation "split-asym" p-values.}
\usage{
diffregr_pval(y1, y2, x1, x2, beta1, beta2, beta, act1, act2, act,
  compute.evals, method.compquadform, acc, epsabs, epsrel, show.warn, n.perm)
}
\arguments{
\item{y1}{Response vector condition 1.}

\item{y2}{Response vector condition 2.}

\item{x1}{Predictor matrix condition 1.}

\item{x2}{Predictor matrix condition 2.}

\item{beta1}{Regression coefficients condition 1.}

\item{beta2}{Regression coefficients condition 2.}

\item{beta}{Pooled regression coefficients.}

\item{act1}{Active-set condition 1.}

\item{act2}{Active-set condition 2.}

\item{act}{Pooled active-set.}

\item{compute.evals}{Method for computation of weights.}

\item{method.compquadform}{Method to compute distribution function of w-sum-of-chi2.}

\item{acc}{See ?davies.}

\item{epsabs}{See ?imhof.}

\item{epsrel}{See ?imhof.}

\item{show.warn}{Show warnings?}

\item{n.perm}{Number of permutations.}
}
\value{
P-value, test statistic, estimated weights.
}
\description{
Computation "split-asym"/"split-perm" p-values.
}
\author{
n.stadler
}

