% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffnet.R
\name{est2.my.ev2}
\alias{est2.my.ev2}
\title{Weights of sum-w-chi2}
\usage{
est2.my.ev2(sig1, sig2, sig, act1, act2, act, include.mean = FALSE)
}
\arguments{
\item{sig1}{no descr}

\item{sig2}{no descr}

\item{sig}{no descr}

\item{act1}{no descr}

\item{act2}{no descr}

\item{act}{no descr}

\item{include.mean}{no descr}
}
\value{
no descr
}
\description{
Compute weights of sum-w-chi2 (2nd order simplification)
}
\details{
*expansion of W in two directions ("dimf>dimg direction" & "dimf>dimg direction")
*simplified computation of weights is obtained by assuming H0 and that X_u~X_v holds
}
\author{
n.stadler
}
\keyword{internal}

