% Generated by roxygen2 (4.0.2): do not edit by hand
\name{est2.my.ev3.diffregr}
\alias{est2.my.ev3.diffregr}
\title{Compute weights of sum-of-weighted-chi2s}
\usage{
est2.my.ev3.diffregr(y1, y2, x1, x2, beta1, beta2, beta, act1, act2, act)
}
\arguments{
\item{y1}{Response vector sample 1.}

\item{y2}{Response vector sample 2.}

\item{x1}{Predictor matrix sample 1.}

\item{x2}{Predictor matrix sample 2.}

\item{beta1}{MLE (regression coefficients) sample 1.}

\item{beta2}{MLE (regression coefficients) sample 2.}

\item{beta}{Pooled MLE (regression coefficients).}

\item{act1}{Active-set sample 1}

\item{act2}{Active-set sample 2}

\item{act}{Pooled active-set}
}
\value{
Eigenvalues of M, respectively the weights.
}
\description{
Compute weights of sum-of-weighted-chi2s
}
\details{
*'2nd order simplification':
  1) Factor out (1-vi)^(d1+d2) "expansion in dimf>dimg direction (old terminology)"
  2) Factor out (1-mu)^d0
*simplified computation of weights is obtained without further invoking H0, or assuming X_u~X_v
}
\author{
n.stadler
}
\keyword{internal}

