% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffnet.R
\name{est2.my.ev3}
\alias{est2.my.ev3}
\title{Compute weights of sum-of-weighted-chi2s}
\usage{
est2.my.ev3(sig1, sig2, sig, act1, act2, act, include.mean = FALSE)
}
\arguments{
\item{sig1}{MLE (covariance matrix) sample 1}

\item{sig2}{MLE (covariance matrix) sample 2}

\item{sig}{Pooled MLE (covariance matrix)}

\item{act1}{Active-set sample 1}

\item{act2}{Active-set sample 2}

\item{act}{Pooled active-set}

\item{include.mean}{Should the mean be in cluded in the likelihood?}
}
\value{
Eigenvalues of M, respectively the weights.
}
\description{
Compute weights of sum-of-weighted-chi2s
}
\details{
*'2nd order simplification':
  1) Factor out (1-vi)^(d1+d2) "expansion in dimf>dimg direction (old terminology)"
  2) Factor out (1-mu)^d0
*simplified computation of weights is obtained without further invoking H0, or assuming X_u~X_v
}
\author{
n.stadler
}
\keyword{internal}

