% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffregr.R
\name{est2.ww.mat.diffregr}
\alias{est2.ww.mat.diffregr}
\title{Estimate weights}
\usage{
est2.ww.mat.diffregr(y1, y2, x1, x2, beta1, beta2, beta, act1, act2, act)
}
\arguments{
\item{y1}{no descr}

\item{y2}{no descr}

\item{x1}{no descr}

\item{x2}{no descr}

\item{beta1}{no descr}

\item{beta2}{no descr}

\item{beta}{no descr}

\item{act1}{no descr}

\item{act2}{no descr}

\item{act}{no descr}
}
\value{
no descr
}
\description{
Estimate weights
}
\details{
estimate W-matrix (using plug-in estimates of Beta-matrix); calculate eigenvalues(W-matrix)
}
\author{
n.stadler
}
\keyword{internal}

