% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mixglasso.R
\name{func.uinit}
\alias{func.uinit}
\title{Initialization of MixGLasso}
\usage{
func.uinit(x, n.comp, init = "kmeans", my.cl = NULL, nstart.kmeans = 1,
  iter.max.kmeans = 10, modelname.hc = "EII")
}
\arguments{
\item{x}{Observed data}

\item{n.comp}{Number of mixture components}

\item{init}{Method used for initialization
init={'cl.init','r.means','random','kmeans','kmeans.hc','hc'}}

\item{my.cl}{Initial cluster assignments;
need to be provided if init='cl.init' (otherwise this param is ignored)}

\item{nstart.kmeans}{Number of random starts in kmeans; default=1}

\item{iter.max.kmeans}{Maximal number of iteration in kmeans; default=10}

\item{modelname.hc}{Model class used in hc; default='EII'}
}
\value{
a list consisting of
\item{u}{responsibilities}
}
\description{
Initialization of responsibilities
}
\author{
n.stadler
}
\keyword{internal}

