% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/netsim.R
\name{getinvcov}
\alias{getinvcov}
\title{Generate an inverse covariance matrix with a given sparsity and
dimensionality}
\usage{
getinvcov(p, s, a.diff = 5, b.diff = 5, magn.diag = 0, emin = 0.1)
}
\arguments{
\item{p}{Dimensionality}

\item{s}{Sparsity}

\item{a.diff}{binomial parameter}

\item{b.diff}{binomial parameter}

\item{magn.diag}{Magnitude}

\item{emin}{e min}
}
\value{
Inverse covariance matrix
Internal function
}
\description{
Generate an inverse covariance matrix with a given sparsity and
dimensionality
}
\keyword{internal}

