% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggmgsa.R
\name{ggmgsa_singlesplit}
\alias{ggmgsa_singlesplit}
\title{Single-split GGMGSA}
\usage{
ggmgsa_singlesplit(x1, x2, gene.sets, gene.names, method.p.adjust = "fdr",
  verbose = TRUE, ...)
}
\arguments{
\item{x1}{centered (scaled) data for condition 1}

\item{x2}{centered (scaled) data for condition 2}

\item{gene.sets}{List of gene-sets.}

\item{gene.names}{Gene names. Each column in x1 (and x2) corresponds
to a gene.}

\item{method.p.adjust}{Method for p-value adjustment (default='fdr').}

\item{verbose}{If TRUE, show output progess.}

\item{...}{Other arguments (see diffnet_singlesplit).}
}
\value{
List of results.
}
\description{
Single-split GGMGSA
}
\author{
n.stadler
}

