% Generated by roxygen2 (4.0.2): do not edit by hand
\name{glasso.invcov}
\alias{glasso.invcov}
\title{Graphical Lasso based on inverse correlation penalty}
\usage{
glasso.invcov(s, rho, penalize.diagonal, term = 10^{     -3 })
}
\arguments{
\item{s}{no descr}

\item{rho}{no descr}

\item{penalize.diagonal}{no descr}

\item{term}{no descr}
}
\value{
w; wi; iter
}
\description{
Graphical Lasso based on inverse correlation penalty
}
\author{
n.stadler
}
\keyword{internal}

