% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggmgsa.R
\name{gsea.iriz.shift}
\alias{gsea.iriz.shift}
\title{Irizarry approach (shift only)}
\usage{
gsea.iriz.shift(x1, x2, gene.sets, gene.names, gs.names = NULL,
  method.p.adjust = "fdr", alternative = "two-sided")
}
\arguments{
\item{x1}{no descr}

\item{x2}{no descr}

\item{gene.sets}{no descr}

\item{gene.names}{no descr}

\item{gs.names}{no descr}

\item{method.p.adjust}{no descr}

\item{alternative}{no descr}
}
\value{
no descr
}
\description{
Irizarry approach (shift only)
}
\author{
n.stadler
}
\keyword{internal}

