% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggmgsa.R
\name{gsea.t2cov}
\alias{gsea.t2cov}
\title{GSA using T2cov-test}
\usage{
gsea.t2cov(x1, x2, gene.sets, gene.names, gs.names = NULL,
  method = "t2cov.lr", method.p.adjust = "fdr")
}
\arguments{
\item{x1}{expression matrix (condition 1)}

\item{x2}{expression matrix (condition 2)}

\item{gene.sets}{list of gene-sets}

\item{gene.names}{gene names}

\item{gs.names}{gene-set names}

\item{method}{method for testing equality of covariance matrices}

\item{method.p.adjust}{method for p-value adjustment (default: 'fdr')}
}
\value{
list of results
}
\description{
GSA using T2cov-test
}
\author{
n.stadler
}
\keyword{internal}

