% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffnet.R
\name{hugepath}
\alias{hugepath}
\title{Graphical Lasso path with huge package}
\usage{
hugepath(s, rholist, penalize.diagonal = NULL, trace = NULL)
}
\arguments{
\item{s}{no descr}

\item{rholist}{no descr}

\item{penalize.diagonal}{no descr}

\item{trace}{no descr}
}
\value{
no descr
}
\description{
Graphical Lasso path with huge package
}
\author{
n.stadler
}
\keyword{internal}

