% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffnet.R
\name{inf.mat}
\alias{inf.mat}
\title{Information Matrix of Gaussian Graphical Model}
\usage{
inf.mat(Sig, include.mean = FALSE)
}
\arguments{
\item{Sig}{Sig=solve(SigInv) true covariance matrix under H0}

\item{include.mean}{no descr}
}
\value{
no descr
}
\description{
Compute Information Matrix of Gaussian Graphical Model
}
\details{
computes E_0[s(Y;Omega)s(Y;Omega)'] where s(Y;Omega)=(d/dOmega) LogLik
}
\author{
n.stadler
}
\keyword{internal}

