% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffnet.R
\name{lambdagrid_mult}
\alias{lambdagrid_mult}
\title{Lambda-grid}
\usage{
lambdagrid_mult(lambda.min, lambda.max, nr.gridpoints)
}
\arguments{
\item{lambda.min}{no descr}

\item{lambda.max}{no descr}

\item{nr.gridpoints}{no descr}
}
\value{
no descr
}
\description{
Lambda-grid (log scale)
}
\author{
n.stadler
}
\keyword{internal}

