% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffnet.R
\name{make_grid}
\alias{make_grid}
\title{Make grid}
\usage{
make_grid(lambda.min, lambda.max, nr.gridpoints, method = "lambdagrid_mult")
}
\arguments{
\item{lambda.min}{no descr}

\item{lambda.max}{no descr}

\item{nr.gridpoints}{no descr}

\item{method}{no descr}
}
\value{
no descr
}
\description{
Make grid
}
\author{
n.stadler
}
\keyword{internal}

