% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffnet.R
\name{mcov}
\alias{mcov}
\title{Compute covariance matrix}
\usage{
mcov(x, include.mean, covMethod = "ML")
}
\arguments{
\item{x}{no descr}

\item{include.mean}{no descr}

\item{covMethod}{no descr}
}
\value{
no descr
}
\description{
Compute covariance matrix
}
\author{
n.stadler
}
\keyword{internal}

