% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffnet.R
\name{mle.ggm}
\alias{mle.ggm}
\title{MLE in GGM}
\usage{
mle.ggm(x, wi, algorithm = "glasso_rho0", rho = NULL, include.mean)
}
\arguments{
\item{x}{no descr}

\item{wi}{no descr}

\item{algorithm}{no descr}

\item{rho}{no descr}

\item{include.mean}{no descr}
}
\value{
no descr
}
\description{
MLE in GGM
}
\author{
n.stadler
}
\keyword{internal}

