% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mixglasso.R
\name{MStepGlasso}
\alias{MStepGlasso}
\title{MStep of MixGLasso}
\usage{
MStepGlasso(x, chromosome = NULL, u, v = NULL, lambda, gamma, pen,
  penalize.diagonal, equal.prob.trans = NULL, term, model = "hmm")
}
\arguments{
\item{x}{no descr}

\item{chromosome}{no descr}

\item{u}{no descr}

\item{v}{no descr}

\item{lambda}{no descr}

\item{gamma}{no descr}

\item{pen}{no descr}

\item{penalize.diagonal}{no descr}

\item{equal.prob.trans}{no descr}

\item{term}{no descr}

\item{model}{no descr}
}
\value{
list consisting of mix.prob, Mu, Sig, SigInv
}
\description{
MStep of MixGLasso
}
\author{
n.stadler
}
\keyword{internal}

