% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffregr.R
\name{my.ev2.diffregr}
\alias{my.ev2.diffregr}
\title{Computation eigenvalues}
\usage{
my.ev2.diffregr(Sig, act, act1, act2)
}
\arguments{
\item{Sig}{no descr}

\item{act}{no descr}

\item{act1}{no descr}

\item{act2}{no descr}
}
\value{
no descr
}
\description{
Computation eigenvalues
}
\author{
n.stadler
}
\keyword{internal}

