% Generated by roxygen2 (4.0.2): do not edit by hand
\name{my.p.adjust}
\alias{my.p.adjust}
\title{P-value adjustment}
\usage{
my.p.adjust(p, method = "fdr")
}
\arguments{
\item{p}{Vector of p-values.}

\item{method}{Method for p-value adjustment (default='fdr').}
}
\value{
Vector of adjusted p-values.
}
\description{
P-value adjustment
}
\author{
n.stadler
}
\keyword{internal}

