% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggmgsa.R
\name{my.ttest}
\alias{my.ttest}
\title{T-test}
\usage{
my.ttest(x1, x2)
}
\arguments{
\item{x1}{no descr}

\item{x2}{no descr}
}
\value{
no descr
}
\description{
T-test (equal variances)
}
\author{
n.stadler
}
\keyword{internal}

