% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffnet.R
\name{mytrunc.method}
\alias{mytrunc.method}
\title{Additional thresholding}
\usage{
mytrunc.method(n, wi, method = "linear.growth", trunc.k = 5)
}
\arguments{
\item{n}{no descr}

\item{wi}{no descr}

\item{method}{no descr}

\item{trunc.k}{no descr}
}
\value{
no descr
}
\description{
Additional thresholding
}
\author{
n.stadler
}
\keyword{internal}

