% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nethet.R
\docType{package}
\name{NetHet-package}
\alias{NetHet-package}
\title{NetHet-package}
\description{
A bioconductor package for high-dimensional exploration of biological network heterogeneity
}
\details{
Includes:
  *Network-based clustering (MixGLasso)
  *Differential network (DiffNet)
  *Differential regression (DiffRegr)
  *Gene-set analysis based on graphical models (GGMGSA)
  *Plotting functions for exploring network heterogeneity
}
\references{
St{\"a}dler, N. and Mukherjee, S. (2013). Two-Sample Testing in High-Dimensional Models.
Preprint \url{http://arxiv.org/abs/1210.4584}.
}

