% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffregr.R
\name{perm.diffregr_pval}
\alias{perm.diffregr_pval}
\title{Computation "split-perm" p-value.}
\usage{
perm.diffregr_pval(y1, y2, x1, x2, act1, act2, act, n.perm)
}
\arguments{
\item{y1}{Response vector condition 1.}

\item{y2}{Response vector condition 2.}

\item{x1}{Predictor matrix condition 1.}

\item{x2}{Predictor matrix condition 2.}

\item{act1}{Active-set condition 1.}

\item{act2}{Active-set condition 2.}

\item{act}{Pooled active-set.}

\item{n.perm}{Number of permutations.}
}
\value{
Permutation based p-value.
}
\description{
Computation "split-perm" p-value.
}
\author{
n.stadler
}
\keyword{internal}

