% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffregr.R
\name{perm.diffregr_teststat}
\alias{perm.diffregr_teststat}
\title{Auxiliary function for computation of "split-perm" p-value.}
\usage{
perm.diffregr_teststat(y1, y2, y12, x1, x2, x12)
}
\arguments{
\item{y1}{Response vector condition 1.}

\item{y2}{Response vector condition 2.}

\item{y12}{Pooled response vector.}

\item{x1}{Predictor matrix condition 1.}

\item{x2}{Predictor matrix condition 2.}

\item{x12}{Pooled predictor matrix}
}
\value{
Test statistic (log-likelihood-ratio statistic).
}
\description{
Auxiliary function for computation of "split-perm" p-value.
}
\author{
n.stadler
}
\keyword{internal}

