% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/netplot.R
\name{plot.diffnet}
\alias{plot.diffnet}
\title{Plotting function for object of class 'diffnet'}
\usage{
\method{plot}{diffnet}(x, ...)
}
\arguments{
\item{x}{object of class 'diffnet'}

\item{...}{Further arguments.}
}
\value{
Histogram over multi-split p-values.
}
\description{
Plotting function for object of class 'diffnet'
}
\author{
nicolas
}

