% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/netplot.R
\name{plot.diffregr}
\alias{plot.diffregr}
\title{Plotting function for object of class 'diffregr'}
\usage{
\method{plot}{diffregr}(x, ...)
}
\arguments{
\item{x}{object of class 'diffregr'}

\item{...}{Further arguments.}
}
\value{
Histogram over multi-split p-values.
}
\description{
Plotting function for object of class 'diffregr'
}
\author{
nicolas
}

