% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.ggmgsa}
\alias{plot.ggmgsa}
\title{Plotting function for object of class 'ggmgmsa'}
\usage{
\method{plot}{ggmgsa}(x, ...)
}
\arguments{
\item{x}{object of class 'ggmgsa'}

\item{...}{Further arguments.}
}
\value{
Boxplot of single-split p-values.
}
\description{
Plotting function for object of class 'ggmgsa'
}
\author{
nicolas
}

