% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotCV}
\alias{plotCV}
\title{plotCV}
\usage{
plotCV(lambda, cv, cv.error, se = TRUE, type = "b", ...)
}
\arguments{
\item{lambda}{no descr}

\item{cv}{no descr}

\item{cv.error}{no descr}

\item{se}{no descr}

\item{type}{no descr}

\item{...}{no descr}
}
\value{
no descr
}
\description{
plotCV
}
\author{
n.stadler
}
\keyword{internal}

