% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.nethetsummary}
\alias{print.nethetsummary}
\title{Print function for object of class 'nethetsummmary'}
\usage{
\method{print}{nethetsummary}(x, ...)
}
\arguments{
\item{x}{object of class 'nethetsummary'}

\item{...}{Other arguments}
}
\value{
Function does not return anything.
}
\description{
Print function for object of class 'nethetsummary'
}
\author{
frankd
}

