% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffregr.R
\name{q.matrix.diffregr}
\alias{q.matrix.diffregr}
\title{Computation Q matrix}
\usage{
q.matrix.diffregr(Sig, a, b, s)
}
\arguments{
\item{Sig}{no descr}

\item{a}{no descr}

\item{b}{no descr}

\item{s}{no descr}
}
\value{
no descr
}
\description{
Computation Q matrix
}
\author{
n.stadler
}
\keyword{internal}

