% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffregr.R
\name{q.matrix.diffregr3}
\alias{q.matrix.diffregr3}
\title{Computation Q matrix}
\usage{
q.matrix.diffregr3(beta.a, beta.b, beta, sig.a, sig.b, sig, Sig, act.a, act.b,
  ss)
}
\arguments{
\item{beta.a}{no descr}

\item{beta.b}{no descr}

\item{beta}{no descr}

\item{sig.a}{no descr}

\item{sig.b}{no descr}

\item{sig}{no descr}

\item{Sig}{no descr}

\item{act.a}{no descr}

\item{act.b}{no descr}

\item{ss}{no descr}
}
\value{
no descr
}
\description{
Computation Q matrix
}
\author{
n.stadler
}
\keyword{internal}

