% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffregr.R
\name{q.matrix.diffregr4}
\alias{q.matrix.diffregr4}
\title{Computation Q matrix}
\usage{
q.matrix.diffregr4(b.mat, act.a, act.b, ss)
}
\arguments{
\item{b.mat}{no descr}

\item{act.a}{no descr}

\item{act.b}{no descr}

\item{ss}{no descr}
}
\value{
no descr
}
\description{
Computation Q matrix
}
\author{
n.stadler
}
\keyword{internal}

