% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffnet.R
\name{q.matrix3}
\alias{q.matrix3}
\title{Compute Q-matrix}
\usage{
q.matrix3(sig, sig.a, sig.b, act.a, act.b, ss)
}
\arguments{
\item{sig}{no descr}

\item{sig.a}{no descr}

\item{sig.b}{no descr}

\item{act.a}{no descr}

\item{act.b}{no descr}

\item{ss}{no descr}
}
\value{
no descr
}
\description{
Compute Q-matrix
}
\author{
n.stadler
}
\keyword{internal}

