% Generated by roxygen2 (4.0.2): do not edit by hand
\name{q.matrix4}
\alias{q.matrix4}
\title{q.matrix4}
\usage{
q.matrix4(b.mat, act.a, act.b, ss)
}
\arguments{
\item{b.mat}{no descr}

\item{act.a}{no descr}

\item{act.b}{no descr}

\item{ss}{no descr}
}
\value{
no descr
}
\description{
q.matrix4
}
\author{
n.stadler
}
\keyword{internal}

