% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffregr.R
\name{screen_cvfix.lasso}
\alias{screen_cvfix.lasso}
\title{Cross-validated Lasso screening and upper bound on number of predictors.}
\usage{
screen_cvfix.lasso(x, y, no.predictors = 10)
}
\arguments{
\item{x}{Predictor matrix.}

\item{y}{Response vector.}

\item{no.predictors}{Upper bound on number of active predictors,}
}
\value{
Active-set.
}
\description{
Cross-validated Lasso screening and upper bound on number of predictors
}
\details{
Computes Lasso coefficients (cross-validation optimal lambda). Truncates
smalles coefficients to zero such that there are no more than no.predictors
non-zero coefficients
}
\examples{
screen_cvfix.lasso(matrix(rnorm(5000),50,100),rnorm(50))
}
\author{
n.stadler
}

