% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffregr.R
\name{screen_cvsqrt.lasso}
\alias{screen_cvsqrt.lasso}
\title{Cross-validated Lasso screening and sqrt-truncation.}
\usage{
screen_cvsqrt.lasso(x, y)
}
\arguments{
\item{x}{Predictor matrix.}

\item{y}{Response vector.}
}
\value{
Active-set.
}
\description{
Cross-validated Lasso screening and sqrt-truncation.
}
\details{
Computes Lasso coefficients (cross-validation optimal lambda). Truncates
smallest coefficients to zero, such that there are no more than sqrt(n)
non-zero coefficients.
}
\examples{
screen_cvsqrt.lasso(matrix(rnorm(5000),50,100),rnorm(50))
}
\author{
n.stadler
}

