% Generated by roxygen2 (4.0.2): do not edit by hand
\name{screen_cvtrunc.lasso}
\alias{screen_cvtrunc.lasso}
\title{Cross-validated Lasso screening and additional truncation.}
\usage{
screen_cvtrunc.lasso(x, y, k.trunc = 5)
}
\arguments{
\item{x}{Predictor matrix.}

\item{y}{Response vector.}

\item{k.trunc}{Truncation constant="number of samples per predictor" (default=5).}
}
\value{
Active-set.
}
\description{
Cross-validated Lasso screening and additional truncation.
}
\details{
Computes Lasso coefficients (cross-validation optimal lambda). Truncates
smallest coefficients to zero, such that there are no more than n/k.trunc
non-zero coefficients.
}
\examples{
screen_cvtrunc.lasso(matrix(rnorm(5000),50,100),rnorm(50))
}
\author{
n.stadler
}

