% Generated by roxygen2 (4.0.2): do not edit by hand
\name{screen_full}
\alias{screen_full}
\title{Screen_full}
\usage{
screen_full(x, include.mean = NULL, length.lambda = NULL,
  trunc.method = NULL, trunc.k = NULL)
}
\arguments{
\item{x}{no descr}

\item{include.mean}{no descr}

\item{length.lambda}{no descr}

\item{trunc.method}{no descr}

\item{trunc.k}{no descr}
}
\value{
no descr
}
\description{
Screen_full
}
\author{
n.stadler
}
\keyword{internal}

