% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffnet.R
\name{screen_lasso}
\alias{screen_lasso}
\title{Screen_lasso}
\usage{
screen_lasso(x, include.mean = NULL, trunc.method = "linear.growth",
  trunc.k = 5)
}
\arguments{
\item{x}{no descr}

\item{include.mean}{no descr}

\item{trunc.method}{no descr}

\item{trunc.k}{no descr}
}
\value{
no descr
}
\description{
Screen_lasso
}
\author{
n.stadler
}
\keyword{internal}

