% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxfunc.R
\name{summary.diffnet}
\alias{summary.diffnet}
\title{Summary function for object of class 'diffnet'}
\usage{
\method{summary}{diffnet}(object, ...)
}
\arguments{
\item{object}{object of class 'diffnet'}

\item{...}{Other arguments.}
}
\value{
aggregated p-values
}
\description{
Summary function for object of class 'diffnet'
}
\author{
nicolas
}

