% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.diffregr}
\alias{summary.diffregr}
\title{Summary function for object of class 'diffregr'}
\usage{
\method{summary}{diffregr}(object, ...)
}
\arguments{
\item{object}{object of class 'diffregr}

\item{...}{Other arguments}
}
\value{
aggregated p-values
}
\description{
Summary function for object of class 'diffregr'
}
\author{
nicolas
}

