% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.ggmgsa}
\alias{summary.ggmgsa}
\title{Summary function for object of class 'ggmgsa'}
\usage{
\method{summary}{ggmgsa}(object, ...)
}
\arguments{
\item{object}{object of class 'ggmgsa'}

\item{...}{Other arguments}
}
\value{
aggregated p-values
}
\description{
Summary function for object of class 'ggmgsa'
}
\author{
nicolas
}

