% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggmgsa.R
\name{t2cov.lr}
\alias{t2cov.lr}
\title{Classical likelihood-ratio test}
\usage{
t2cov.lr(x1, x2, include.mean = FALSE)
}
\arguments{
\item{x1}{no descr}

\item{x2}{no descr}

\item{include.mean}{no descr}
}
\value{
no descr
}
\description{
Classical likelihood-ratio test (equality of covariance matrices)
}
\author{
n.stadler
}
\keyword{internal}

