% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggmgsa.R
\name{t2diagcov.lr}
\alias{t2diagcov.lr}
\title{Diagonal-restricted likelihood-ratio test}
\usage{
t2diagcov.lr(x1, x2, include.mean = FALSE)
}
\arguments{
\item{x1}{no descr}

\item{x2}{no descr}

\item{include.mean}{no descr}
}
\value{
no descr
}
\description{
Diagonal-restricted likelihood-ratio test
}
\author{
n.stadler
}
\keyword{internal}

