% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggmgsa.R
\name{test.t2}
\alias{test.t2}
\title{HotellingsT2}
\usage{
test.t2(x1, x2)
}
\arguments{
\item{x1}{no descr}

\item{x2}{no descr}
}
\value{
no descr
}
\description{
HotellingsT2
}
\author{
n.stadler
}
\keyword{internal}

