% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mixglasso.R
\name{tr}
\alias{tr}
\title{Compute trace of matrix}
\usage{
tr(m)
}
\arguments{
\item{m}{no descr}
}
\value{
trace of matrix
}
\description{
Compute trace of matrix
}
\author{
n.stadler
}
\keyword{internal}

