% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffregr.R
\name{twosample_single_regr}
\alias{twosample_single_regr}
\title{old single-split function for diffregr}
\usage{
twosample_single_regr(y1, y2, x1, x2, n.screen.pop1 = 100,
  n.screen.pop2 = 100, screen.meth = "screen_cvmin.lasso",
  compute.evals = "est2.my.ev3.diffregr")
}
\arguments{
\item{y1}{no descr}

\item{y2}{no descr}

\item{x1}{no descr}

\item{x2}{no descr}

\item{n.screen.pop1}{no descr}

\item{n.screen.pop2}{no descr}

\item{screen.meth}{no descr}

\item{compute.evals}{no descr}
}
\value{
no descr
}
\description{
Old single-split function for diffregr
}
\author{
n.stadler
}
\keyword{internal}

