% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ww.mat2}
\alias{ww.mat2}
\title{Calculates eigenvalues of weight-matrix (using 1st order simplification)}
\usage{
ww.mat2(imat, act, act1, act2)
}
\arguments{
\item{imat}{no descr}

\item{act}{I_uv}

\item{act1}{I_u}

\item{act2}{I_v}
}
\value{
no descr
}
\description{
Calculates eigenvalues of weight-matrix (using 1st order simplification)
}
\details{
calculation based on true information matrix
}
\author{
n.stadler
}
\keyword{internal}

