% Generated by roxygen2 (4.0.2): do not edit by hand
\name{beta.mat}
\alias{beta.mat}
\title{Compute beta-matrix}
\usage{
beta.mat(ind1, ind2, sig1, sig2, sig)
}
\arguments{
\item{ind1}{no descr}

\item{ind2}{no descr}

\item{sig1}{no descr}

\item{sig2}{no descr}

\item{sig}{no descr}
}
\value{
no descr
}
\description{
Compute beta-matrix
}
\details{
beta-matrix=E[s_ind1(Y;sig1) s_ind2(Y;sig2)'|sig]
}
\author{
n.stadler
}
\keyword{internal}

