% Generated by roxygen2 (4.0.2): do not edit by hand
\name{error.bars}
\alias{error.bars}
\title{Error bars for plotCV}
\usage{
error.bars(x, upper, lower, width = 0.02, ...)
}
\arguments{
\item{x}{no descr}

\item{upper}{no descr}

\item{lower}{no descr}

\item{width}{no descr}

\item{...}{no descr}
}
\value{
no descr
}
\description{
Error bars for plotCV
}
\author{
n.stadler
}
\keyword{internal}

