% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mixglasso.R
\name{loglik_mix}
\alias{loglik_mix}
\title{Log-likelihood for mixture model}
\usage{
loglik_mix(x, mix.prob, Mu, Sig)
}
\arguments{
\item{x}{no descr}

\item{mix.prob}{no descr}

\item{Mu}{no descr}

\item{Sig}{no descr}
}
\value{
log-likelihood
}
\description{
Log-likelihood for mixture model
}
\author{
n.stadler
}
\keyword{internal}

