% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxfunc.R
\name{summary.diffregr}
\alias{summary.diffregr}
\title{Summary function for object of class 'diffregr'}
\usage{
\method{summary}{diffregr}(object, ...)
}
\arguments{
\item{object}{object of class 'diffregr}

\item{...}{Other arguments}
}
\value{
aggregated p-values
}
\description{
Summary function for object of class 'diffregr'
}
\author{
nicolas
}

