% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mixglasso.R
\name{sumoffdiag}
\alias{sumoffdiag}
\title{Sum of non-diag elements of a matrix}
\usage{
sumoffdiag(m)
}
\arguments{
\item{m}{no descr}
}
\value{
Sum of non-diag elements
}
\description{
Sum of non-diag elements of a matrix
}
\author{
n.stadler
}
\keyword{internal}

