% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffnet.R
\name{ww.mat}
\alias{ww.mat}
\title{Weight-matrix and eigenvalues}
\usage{
ww.mat(imat, act, act1, act2)
}
\arguments{
\item{imat}{no descr}

\item{act}{I_uv}

\item{act1}{I_u}

\item{act2}{I_v}
}
\value{
no descr
}
\description{
Calculates weight-matrix and eigenvalues
}
\details{
calculation based on true information matrix
}
\author{
n.stadler
}
\keyword{internal}

